ActivemqConsumer
Consumer läuft
