--- Consumer/src/main/java/activemq/ActiveConsumer.java
@@ -1,50 +1,0 @@
-package kafka;
-
-/**
- * Created by Uno on 15.10.2016.
- */
-    public static class ActiveConsumer implements Runnable, ExceptionListener {
-        public void run() {
-            try {
-
-                // Create a ConnectionFactory
-                ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory("//IP:PORT");
-
-                // Create a Connection
-                Connection connection = connectionFactory.createConnection();
-                connection.start();
-
-                connection.setExceptionListener(this);
-
-                // Create a Session
-                Session session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-
-                // Create the destination (Topic)
-                Destination destination = session.createTopic("prod");
-
-                // Create a MessageConsumer from the Session to the Topic
-                MessageConsumer consumer = session.createConsumer(destination);
-
-                // Wait for a message
-                Message message = consumer.receive(1000);
-
-                if (message instanceof TextMessage) {
-                    TextMessage textMessage = (TextMessage) message;
-                    String text = textMessage.getText();
-                    System.out.println("Received: " + text);
-                } else {
-                    System.out.println("Received: " + message);
-                }
-
-                consumer.close();
-                session.close();
-                connection.close();
-            } catch (Exception e) {
-                System.out.println("Caught: " + e);
-                e.printStackTrace();
-            }
-        public static void main(String args[]){
-            ActiveConsumer a = new ActiveConsumer();
-            a.run();
-        }
-    }
