+++ Consumer/src/main/java/main/Main.java
@@ -14,9 +14,9 @@ public class Main {
     public static MyMachine myMachine;
 
     public static void main (String[] args){
-        Consumer consumer = new Consumer("192.168.99.100:1001", "prod");
-        myMachine = new MyMachine();
-        consumer.start();
+        //Consumer consumer = new Consumer("192.168.99.100:1001", "prod");
+       // myMachine = new MyMachine();
+      // consumer.start();
 
     }
 }
